requirements.txt
-e git+git@github.com:ryanlow-tw/flask-app-2.git@2a34be3df3f2362bdc1a66595d92ab69bc9dfc7c#egg=app
attrs==21.2.0
click==8.0.1
Flask==2.0.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pymongo==3.12.0
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.19.0
toml==0.10.2
Werkzeug==2.0.1
